Add separate public export to config package so that it's easier to reexport public types
